Refactor utility components
